pyproject.toml
[project]
name = "imdb-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.68",
    "jupyter>=1.1.1",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]
